vuforia game objects as buttons
This is a unity project that uses game objects as buttons. The script
that made this possible is the 'color changer' script. Its a terrible
name for what it does, because it doesnt actually change any colors. it
just makes the UI show up when you click on the ar object. the script is
attached to the game object itself.